+++ CHANGELOG.rst
@@ -16,6 +16,7 @@ Unreleased
 * add `workspace_bodies` to the `get_collision_fn` to check collisions with the obstacles specified in a URDF file.
 * move `ik_interface` module from application side to this repo, since it's "universal" for fixed-end robot. Might need to add a separete one for robots with moving base later.
 * enable travis ci unit test, collision_fn well tested
+* `get_floating_body_collision_fn` to check a body without joints's collision. Associated test added.
 
 **Changed**